🎨 feat(assets): add dark and light versions of appily icon The commit adds two new SVG files, "appily-icon-dark.svg" and "appily-icon-light.svg", which contain the dark and light versions of the appily icon respectively. These new assets provide flexibility in using the appropriate version of the icon based on the design requirements or theme of the application.
🎉 feat(appily-logo-dark.svg): add dark version of Appily logo
The commit adds a new file, "appily-logo-dark.svg", which contains the dark version of the Appily logo. This allows for more flexibility in using the logo in different contexts, such as on a dark background or in a dark-themed application.

🎉 feat(appily-logo-light.svg): add new Appily logo in light mode
The commit adds a new SVG file named "appily-logo-light.svg" which contains the code for the Appily logo in light mode. This logo can be used in the application to represent the brand in a visually appealing way.

📦 chore(favicon.svg): add favicon.svg
📦 chore(Icon.tsx): add Icon component
📦 chore(Logo.tsx): add Logo component
📦 chore(payload.config.ts): update admin meta and components
The changes include adding a new favicon.svg file, as well as adding the Icon and Logo components to the admin components in the payload.config.ts file. The admin meta in the payload.config.ts file is also updated to include a title suffix, favicon, and ogImage.