🔧 chore(.gitignore): add /dump directory to .gitignore 🔧 chore(CarouselCards.ts): add "versions" field with "drafts" option set to true 🔧 chore(payload-types.ts): remove unused payment related fields from Form and FormSubmission interfaces, update Partner interface
The /dump directory is now added to the .gitignore file to exclude it from version control. This is done to prevent accidental commits of sensitive data.

In the CarouselCards.ts file, a new "versions" field is added to the CarouselCards collection configuration. The "drafts" option is set to true, indicating that the collection supports draft versions.

In the payload-types.ts file, the unused payment related fields in the Form and FormSubmission interfaces are removed. Additionally, the Partner interface is updated to make the "acroynm" field optional.